fix: downgrading default notebook platform
Beginning of 2023 sagemaker stopped supporting the notebooks on al1 platform.
Now by default all notebooks created on sagemaker use the notebook-al2-v2 Platform.
This broke our current notebook template where we stopped getting the environment variables in the notebook.

This fix downgrades the platform to notebook-al2-v1 platform to fix the problem.

Tested it with a pipeline I created and it worked.