fix: Metadata component loading
This fixes an issue when rendering the ControlledMetadata component when
there is no CSV schema in S3. A missing CSV would cause the request to
throw and there was no error-handling to continue the render if it
failed. There were similar error-handling problems when an asset had
inconsistent metadata, or no schema.

Other changes:
- Removed some debugging console logs.
- Added a type assertion for AxiosErrors to make error-handling easier.