Fix for metadata race condition
The ControlledMetadata component was setting the state of the valid
fields directly, which worked fine when entering the values, however if
the user navigated to the next step in the wizard and then back, the
rows would fight over the current value of the validity array, resulting
in a race condition. Switching the state setter to use a function
instead of a value fixes this race condition.